+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Fixed
 
+- Accessing json properties from app.config inside a view looked up view.config instead of app.config (https://github.com/edvin/tornadofx/issues/346)
+
 ### Changes